# Prefer Hash#[] over Set#.include? for speed