@schneems 2016-06-15
Actually looks like we can go even faster with case statements

```
Comparison:
                case:  6852426.1 i/s
                hash:  3622063.5 i/s - 1.89x slower
                set :  3400254.4 i/s - 2.02x slower
```