@jeremy 2016-06-15
Was just writing to suggest `case`, since Ruby gives us a helping hand 😁

Wonder whether `validate_processor_result!` is over-defensive, anyway, like running a linter in production code. Are a processor's results better validated by its test suite?