@schneems 2016-06-16
We need some kind of way to enable a "debug" or "strict" mode when it comes to these checks. They're not even really that useful for debugging, they only guarantee a certain subset of types are used, they don't guarantee the correct types are used in the correct places. I guess the checks are done to blow up before we try to serialize a non-serializable type to the cache by accident.The deep checking of types is really bad for source maps which is an array of hashes with values that hold arrays.

If this is really so expensive, we should definitely look for ways to work around having to do these checks in production.

I tried implementing the logic using a case statement:

```ruby
    def valid_processor_metadata_value?(value)
      case value
      when String, Symbol, Fixnum, Bignum, TrueClass, FalseClass, NilClass
        true
      when Set, Array
        value.all? { |v| valid_processor_metadata_value?(v) }
      when Hash
        value.all? do |(key, value)|
          valid_processor_metadata_value?(key) &&
            valid_processor_metadata_value?(value)
        end
      else
        false
      end
    end
```

Then I set up my system to run `assets:precompile` 50 times and record the results for the codetriage website:

```ruby
task "assets:bench" do
  measure = []

  50.times do
    measure << Benchmark.measure do
      `rm -rf tmp/cache/assets/sprockets/v4.0/ ; rm -rf public/assets; time RAILS_ENV=production bundle exec rake assets:precompile`
    end.real
  end
  puts "================ DONE ================"
  puts measure.join("\n")
end
```

I put the results in a spreadsheet

```
Name       Avg (seconds)   STDev(seconds)
----------------------------------------
Current	13.96280399	  0.3321853933
Hash	12.54463898   0.250917801
Case	13.13691678	  1.628405121
```

The case statement is faster than the "current" but not within the stdev. Hash is the fastest even with stdev. On the average case my assets compile about 10% faster when we switched over to a hash.

I think the case statement might care where the types are in the statement, or maybe it might care how many values in `when` are available for it to compare. Or maybe `===` is slower than `thing.class == String`. I don't know. I'm curious to look into it more tomorrow.

If you're interested in messing around with optimizing the case statement I wrote a script that compares it between hash and set https://gist.github.com/schneems/925ff1988066d4d582915fdf948ca142 for the non "complex" (i.e. Set, Array, Hash) data types.