@schneems 2016-06-17
Looked into it more, still not sure why case is slower. Need to move on. We can revisit some other time.