@tenderlove 2016-06-17
> Or maybe === is slower than thing.class == String. I don't know. I'm curious to look into it more tomorrow.

Yes.  `Class#===(obj)` does an `oby.is_a?(Class)` where `thing.class == String` is only comparing the classes.  IOW switching from a case statement to a hash lookup would mean you don't support subclasses of those classes.  But it looks like that's not a real use case, so I'd just go with the hash.