@vincentwoo 2016-06-21
Whoa, this is a little weird to me. Any thoughts on why `Set.include?` is slower than `Hash#[]`? I thought Set was supposed to be a special-purpose class explicitly to make set operations like `include?` fast.