@schneems 2016-06-21
Calls to `Hash#[]` are optimized by the interpreter. Here's some disassembled code:

```ruby
code = <<-END
  set = Set.new
  set.include?(foo.class)
END
puts RubyVM::InstructionSequence.compile(code).disasm
```

Produces

```
== disasm: #<ISeq:<compiled>@<compiled>>================================
local table (size: 2, argc: 0 [opts: 0, rest: -1, post: 0, block: -1, kw: -1@-1, kwrest: -1])
[ 2] set
0000 trace            1                                               (   1)
0002 getinlinecache   9, <is:0>
0005 getconstant      :Set
0007 setinlinecache   <is:0>
0009 opt_send_without_block <callinfo!mid:new, argc:0, ARGS_SIMPLE>, <callcache>
0012 setlocal_OP__WC__0 2
0014 trace            1                                               (   2)
0016 getlocal_OP__WC__0 2
0018 putself
0019 opt_send_without_block <callinfo!mid:foo, argc:0, FCALL|VCALL|ARGS_SIMPLE>, <callcache>
0022 opt_send_without_block <callinfo!mid:class, argc:0, ARGS_SIMPLE>, <callcache>
0025 opt_send_without_block <callinfo!mid:include?, argc:1, ARGS_SIMPLE>, <callcache>
0028 leave
```

While

```ruby
code = <<-END
  hash = {}
  hash[foo.class]
END
puts RubyVM::InstructionSequence.compile(code).disasm
```

produces

```
== disasm: #<ISeq:<compiled>@<compiled>>================================
local table (size: 2, argc: 0 [opts: 0, rest: -1, post: 0, block: -1, kw: -1@-1, kwrest: -1])
[ 2] hash
0000 trace            1                                               (   1)
0002 newhash          0
0004 setlocal_OP__WC__0 2
0006 trace            1                                               (   2)
0008 getlocal_OP__WC__0 2
0010 putself
0011 opt_send_without_block <callinfo!mid:foo, argc:0, FCALL|VCALL|ARGS_SIMPLE>, <callcache>
0014 opt_send_without_block <callinfo!mid:class, argc:0, ARGS_SIMPLE>, <callcache>
0017 opt_aref         <callinfo!mid:[], argc:1, ARGS_SIMPLE>, <callcache>
```

I think it's that `opt_aref` is faster than `opt_send_without_block`, but IANA(Ruby)C developer.