@vincentwoo 2016-06-21
If this is true, is there ever an appropriate time to use `Set` and not a `Hash` in Ruby?