@schneems 2016-06-22
I think it's supposed to be faster for set operations such as unions and calculating intersections.

I could be wrong but I think set is powered by a hash under the hood.