@schneems 2016-06-22
Looks like Set is written in ruby https://github.com/ruby/ruby/blob/trunk/lib/set.rb and it is powered by a hash. Based on that a Hash will always be faster, however it might not be as convenient.